+++ src/main/java/me/mrCookieSlime/Slimefun/api/EmptyBlockData.java
@@ -20,7 +20,7 @@ class EmptyBlockData extends BlockInfoConfig {
     }
 
     @Override
-    protected void store(String path, Object value) {
+    public void setValue(String path, Object value) {
         throw new UnsupportedOperationException("Cannot store values (" + path + ':' + value + " on a read-only data object!");
     }